finalized t1k processing script

+++ genotypes/genotypeDRB.sh
@@ -3,6 +3,8 @@
 ## Usage: sh genotypeDRB.sh
 ## The fastq files location can be specified in ../control.txt
 ## The sample ID to run can be modified in ../testID.txt
+## Output: a directory 'alleles' with all the T1K result and 'HLAgenotypes.csv' which processed
+##         the result to a format that fits MHConstructor
 
 source ../control.txt
 repoDir=${binDir}/MHConstructor
@@ -12,7 +14,13 @@ set -e
 
 for i in $ids; do
     echo $i
-    ${progT1K} -1 ${targetCapture}/${i}_${R1ext} -2 ${targetCapture}/${i}_${R2ext} --preset hla -f ${repoDir}/tools/T1K/DRB1_IMGT_010324.fasta -o ${i} -t 8 --skipPostAnalysis --od ./alleles
+    R1="$(find ${targetCapture} -name ${i}*${R1ext})"
+    R2="$(find ${targetCapture} -name ${i}*${R2ext})"
+    if [[ -z "$R1" ]]; then
+        continue
+    fi
+
+    ${progT1K} -1 ${R1} -2 ${R2} --preset hla -f ${repoDir}/tools/T1K/DRB1_IMGT_010324.fasta -o ${i} -t 8 --skipPostAnalysis --od ./alleles
     cd ./alleles
     find . ! -name '*_allele.tsv' -type f -exec rm -f {} +
     cd ../
